Remove ZP_ADDR and reusing cells.
When adding to a cell that has not been considered before, it means it has only one neighbour. This means that the priority of the to-be-added cell is anyway maximum you can get, so we might just as well drop it. This, in turn, allowed getting rid of ZP_ADDR completely, as it was now used only in one place.